requirements.txt
python-dotenv  # Manage environment variables
requests       # For API calls
argparse       # Parse command-line arguments
black==24.3.0
flake8==6.0.0
pytest
pytest-watch
toml
pre-commit

